Reorganize test folders
